backend/requirements.txt
# FastAPI and core dependencies
setuptools>=67
wheel>=0.40

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
prisma==0.10.0
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background tasks
celery==5.3.4
redis==5.0.1

numpy==1.26.2  # Changed from 1.24.3 for Python 3.11 compatibility

# File storage
boto3==1.34.0
python-magic==0.4.27

# Payment processing
stripe==7.8.0

# Email
fastapi-mail==1.4.1

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment
python-dotenv==1.0.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80  # Changed to headless version for server